fix: remove DISIR_CONTEXT_TEMPLATE
Remove the traces of DISIR_CONTEXT_TEMPLATE from the codebase.
It was initially introduced when mocking up the library
but should not polute the library if not part of its implementation.

As of now, this feature is not prioritized. In any regard,
it will make the schema versioning much more complex.