fix: plugin_test creates new mold per request.
Previous implementation had pre-allocated molds - which it gave a
reference to on request. This has the bad side-effect of allowing
the user to decref and delete others molds. Therefore, we allocate a
new one per request instead.
Also handle the case where an entry was entered with NULL,
which has the unfortunate side-effect of SIGTERM'ing the entire
map. Spent too many hours on that one than I care to admit.