feat: add dx_validate_context
This method is used to validate any context state and its
children. Its return status depends on the parent constructing state.
If the parent is constructing, all state deeming an input context to
invalid will result in DISIR_STATUS_INVALID_CONTEXT.
If the parent is finalized, the appropriate status codes
indicating the actual error is returned.
This may be used to discriminate between the finalizing state
of input contexts.

It will clear CONTEXT_STATE_INVALID before re-assigning
it should an error be encountered. This is useful
to re-calculate previously deemed invalid contexts
that may be updated, e.g. have their values updated
to be within restriction limitations.